+++ Source/MQTTnet/Server/Events/InterceptingPacketEventArgs.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Collections;
 using System.Threading;
 using MQTTnet.Packets;
 
@@ -10,30 +11,44 @@ namespace MQTTnet.Server
 {
     public sealed class InterceptingPacketEventArgs : EventArgs
     {
+        public InterceptingPacketEventArgs(CancellationToken cancellationToken, string clientId, string endpoint, MqttPacket packet, IDictionary sessionItems)
+        {
+            CancellationToken = cancellationToken;
+            ClientId = clientId ?? throw new ArgumentNullException(nameof(clientId));
+            Endpoint = endpoint;
+            Packet = packet ?? throw new ArgumentNullException(nameof(packet));
+            SessionItems = sessionItems;
+        }
+
         /// <summary>
-        /// Gets the client ID which has sent the packet or will receive the packet.
+        ///     Gets the cancellation token from the connection managing thread.
+        ///     Use this in further event processing.
         /// </summary>
-        public string ClientId { get; internal set; }
-        
+        public CancellationToken CancellationToken { get; }
+
         /// <summary>
-        /// Gets the endpoint of the sending or receiving client.
+        ///     Gets the client ID which has sent the packet or will receive the packet.
         /// </summary>
-        public string Endpoint { get; internal set; }
-        
+        public string ClientId { get; }
+
         /// <summary>
-        /// Gets or sets the MQTT packet which was received or will be sent.
+        ///     Gets the endpoint of the sending or receiving client.
+        /// </summary>
+        public string Endpoint { get; }
+
+        /// <summary>
+        ///     Gets or sets the MQTT packet which was received or will be sent.
         /// </summary>
         public MqttPacket Packet { get; set; }
 
         /// <summary>
-        /// Gets or sets whether the packet should be processed or not.
+        ///     Gets or sets whether the packet should be processed or not.
         /// </summary>
         public bool ProcessPacket { get; set; } = true;
-        
+
         /// <summary>
-        /// Gets the cancellation token from the connection managing thread.
-        /// Use this in further event processing.
+        ///     Gets or sets a key/value collection that can be used to share data within the scope of this session.
         /// </summary>
-        public CancellationToken CancellationToken { get; internal set; }
+        public IDictionary SessionItems { get; }
     }
 }