+++ .github/workflows/ReleaseNotes.md
@@ -4,3 +4,4 @@
 * [Server] Added a new event (_ClientAcknowledgedPublishPacketAsync_) which is fired whenever a client acknowledges a PUBLISH packet (QoS 1 and 2, #487).
 * [Server] Exposed channel adapter (HTTP context etc.) in connection validation (#1125).
 * [Server] The event _InterceptingPublishAsync_ is now also called for injected application messages (#Jeanot-Zubler).
+* [Server] Exposed session items in multipe events (#1291).